refactor: reorganize customer-related files and restore deleted functionality
