+++ backend/controllers/customer.controller.ts
@@ -1,5 +1,5 @@
 import { Context } from "node:vm";
-import { fetchCustomers } from "../services/customer_service.ts";
+import { fetchCustomers } from "../services/customer.service.ts";
 
 export const getCustomers = async (c: Context) => {
     const customers = await fetchCustomers();
