+++ backend/services/customer.service.ts
@@ -1,4 +1,4 @@
-import connection from "../connection/db_connect.ts";
+import connection from "../connection/db.connect.ts";
 
 export async function fetchCustomers() {
     const [rows] = await connection.execute("SELECT * FROM customers");
